requirements.txt
requests
paho-mqtt
notify2